created a new file just for the createContext()
